update providers tests
